change user_model to example (we change schema of database)
