+++ src/main/java/com/spring/chat/example/User.java
@@ -1,0 +1,50 @@
+package com.spring.chat.example;
+
+import io.swagger.annotations.ApiModel;
+
+import javax.persistence.*;
+
+import javax.persistence.Entity;
+
+@ApiModel
+@Entity
+@Table(name="example")
+public class User {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    @Column(unique = true)
+    private String username;
+    private String password;
+
+    public User(Long id, String username, String password) {
+        this.id = id;
+        this.username = username;
+        this.password = password;
+    }
+    public User(){}
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public String getUsername() {
+        return username;
+    }
+
+    public void setUsername(String username) {
+        this.username = username;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
+}
