+++ src/main/java/com/spring/chat/example/UserMapper.java
@@ -1,10 +1,9 @@
-package com.spring.chat.user;
+package com.spring.chat.example;
 
 import org.mapstruct.Mapper;
 
 @Mapper
 public interface UserMapper {
-    FriendDto toFriendDto(UserDto userDto);
     User toUserEntity(UserDto userDto);
     UserDto toUserDto(User user);
 }
