+++ src/main/java/com/spring/chat/example/UserRepo.java
@@ -1,4 +1,4 @@
-package com.spring.chat.user;
+package com.spring.chat.example;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.query.Param;
