+++ src/main/resources/database/patches/01-old/01_18_create_example_table.xml
@@ -3,13 +3,13 @@
         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
         xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-3.1.xsd">
 
-    <changeSet id="01_18_create_user_table" author="alewandowska">
+    <changeSet id="01_18_create_example_table" author="alewandowska">
         <preConditions onFail="MARK_RAN">
             <not>
-                <tableExists tableName="user"/>
+                <tableExists tableName="example"/>
             </not>
         </preConditions>
-        <createTable tableName="user">
+        <createTable tableName="example">
             <column name="id" type="bigint">
                 <constraints primaryKey="true" unique="true"/>  </column>
             <column name="username" type="varchar(100)">
@@ -19,4 +19,9 @@
         </createTable>
     </changeSet>
 
+    <changeSet id="01_20_update_example_table_add_autoincrement" author="alewandowska">
+        <addAutoIncrement tableName="example"
+                          columnDataType="bigint"
+                          columnName="id"/>
+    </changeSet>
 </databaseChangeLog>
