--- src/main/resources/database/patches/01_24_create_friend_table.xml
@@ -1,26 +1,0 @@
-<databaseChangeLog
-        xmlns="http://www.liquibase.org/xml/ns/dbchangelog"
-        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-        xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-3.1.xsd">
-
-    <changeSet id="01_24_create_friend_table" author="alewandowska">
-        <preConditions onFail="MARK_RAN">
-            <not>
-                <tableExists tableName="friend"/>
-            </not>
-        </preConditions>
-        <createTable tableName="friend">
-            <column name="id" type="bigint">
-            <constraints primaryKey="true" unique="true"/>  </column>
-            <column name="friend_id" type="bigint">
-                <constraints nullable="false"/>                 </column>
-            <column name="user_id" type="bigint">
-                <constraints nullable="false"/>                 </column>
-        </createTable>
-    </changeSet>
-
-    <changeSet id="01_24_update_friend_table_add_constraints" author="alewandowska">
-        <addForeignKeyConstraint baseTableName="friend" baseColumnNames="friend_id" referencedTableName="user" referencedColumnNames="id" constraintName="friend_id_fk"/>
-        <addForeignKeyConstraint baseTableName="friend" baseColumnNames="user_id" referencedTableName="user" referencedColumnNames="id" constraintName="user_id_fk"/>
-    </changeSet>
-</databaseChangeLog>
