--- src/main/resources/database/patches/02_01_update_user_table_rename_table.xml
@@ -1,10 +1,0 @@
-<databaseChangeLog
-        xmlns="http://www.liquibase.org/xml/ns/dbchangelog"
-        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-        xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-3.1.xsd">
-
-    <changeSet id="02_01_update_user_table_rename_table" author="alewandowska">
-    <renameTable  newTableName="user_model"
-                  oldTableName="user"/>
-    </changeSet>
-</databaseChangeLog>
